pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-session-recorder"
version = "0.1.0"
description = "A development session recorder that tracks AI interactions, git commits, and file changes"
authors = [{name = "Chandler Hardy"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "click>=8.1.0",
    "sqlalchemy>=2.0.0",
    "gitpython>=3.1.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
chronicle = "backend.main:cli"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
